Revert "Refactoring"
This reverts commit 5c24afae3a971b23f535a2ff824d071630f53f40.
