+++ aws-xray-recorder-sdk-core/src/main/java/com/amazonaws/xray/strategy/DynamicSegmentNamingStrategy.java
@@ -17,14 +17,21 @@ package com.amazonaws.xray.strategy;
 
 import com.amazonaws.xray.entities.SearchPattern;
 import java.util.Optional;
-
 import javax.servlet.http.HttpServletRequest;
+import org.apache.commons.logging.Log;
+import org.apache.commons.logging.LogFactory;
 
 /**
  * @deprecated Use {@link SegmentNamingStrategy#dynamic(String)}.
  */
 @Deprecated
-public class DynamicSegmentNamingStrategy extends com.amazonaws.xray.strategy.interfaces.DynamicSegmentNamingStrategy implements com.amazonaws.xray.strategy.SegmentNamingStrategy {
+public class DynamicSegmentNamingStrategy implements SegmentNamingStrategy {
+    private static final Log logger =
+        LogFactory.getLog(DynamicSegmentNamingStrategy.class);
+
+    private final String recognizedHosts;
+    private final String fallbackName;
+
     /**
      * Creates an instance of {@code DynamicSegmentNamingStrategy} with the provided {@code fallbackName} and a
      * {@code recognizedHosts} value of "*".
@@ -34,11 +41,11 @@ public class DynamicSegmentNamingStrategy extends com.amazonaws.xray.strategy.in
      *  value of the {@code AWS_XRAY_TRACING_NAME} environment variable or {@code com.amazonaws.xray.strategy.tracingName} system
      *  property, if either are set to a non-empty value.
      *
-     * @deprecated Use {@link com.amazonaws.xray.strategy.jakarta.SegmentNamingStrategy#dynamic(String)}.
+     * @deprecated Use {@link SegmentNamingStrategy#dynamic(String)}.
      */
     @Deprecated
     public DynamicSegmentNamingStrategy(String fallbackName) {
-        super(fallbackName, "*");
+        this(fallbackName, "*");
     }
 
     /**
@@ -51,9 +58,9 @@ public class DynamicSegmentNamingStrategy extends com.amazonaws.xray.strategy.in
      *  variable or {@code com.amazonaws.xray.strategy.tracingName} system property, if either are set to a non-empty value.
      * @param recognizedHosts
      *  the pattern to match the incoming host header value against. This pattern is compared against the incoming host header
-     *  using the {@link SearchPattern#wildcardMatch(String, String)} method.
+     *  using the {@link com.amazonaws.xray.entities.SearchPattern#wildcardMatch(String, String)} method.
      *
-     * @deprecated Use {@link com.amazonaws.xray.strategy.jakarta.SegmentNamingStrategy#dynamic(String, String)}.
+     * @deprecated Use {@link SegmentNamingStrategy#dynamic(String, String)}.
      */
     // Instance method is called before the class is initialized. This can cause undefined behavior, e.g., if getOverrideName
     // accesses fallbackName. This class doesn't really need to be exposed to users so we suppress for now and will clean up after
@@ -61,7 +68,21 @@ public class DynamicSegmentNamingStrategy extends com.amazonaws.xray.strategy.in
     @SuppressWarnings("nullness")
     @Deprecated
     public DynamicSegmentNamingStrategy(String fallbackName, String recognizedHosts) {
-        super(fallbackName, recognizedHosts);
+        String overrideName = getOverrideName();
+        if (overrideName != null) {
+            this.fallbackName = getOverrideName();
+            if (logger.isInfoEnabled()) {
+                logger.info("Environment variable " + NAME_OVERRIDE_ENVIRONMENT_VARIABLE_KEY + " or system property "
+                            + NAME_OVERRIDE_SYSTEM_PROPERTY_KEY
+                            + " set. Overriding DynamicSegmentNamingStrategy constructor argument. Segments generated with this "
+                            + "strategy will be named: " + this.fallbackName
+                            + " when the host header is unavilable or does not match the provided recognizedHosts pattern.");
+            }
+        } else {
+            this.fallbackName = fallbackName;
+        }
+
+        this.recognizedHosts = recognizedHosts;
     }
 
     /**
@@ -81,7 +102,7 @@ public class DynamicSegmentNamingStrategy extends com.amazonaws.xray.strategy.in
     public String nameForRequest(HttpServletRequest request) {
         Optional<String> hostHeaderValue = Optional.ofNullable(request.getHeader("Host"));
         if (hostHeaderValue.isPresent() &&
-                (null == recognizedHosts || SearchPattern.wildcardMatch(recognizedHosts, hostHeaderValue.get()))) {
+            (null == recognizedHosts || SearchPattern.wildcardMatch(recognizedHosts, hostHeaderValue.get()))) {
             return hostHeaderValue.get();
         }
         return fallbackName;
